fix: hover actions on GTD tab-list rows
